resolvido o problema das diferenças

+++ TrabIPPD.py
@@ -1,4 +1,3 @@
-print("[*] Carregando bibliotecas...", end='')
 import dropbox
 import os
 import getpass
@@ -7,10 +6,13 @@ import shutil
 import sys
 import base64
 import platform
+import colorama
+from hurry.filesize import size
 from termcolor import cprint
 from Crypto import Random
 from Crypto.Cipher import AES
-cprint("OK", "green")
+
+colorama.init(autoreset=True)
 
 dbx = dropbox.Dropbox("D-O1ZZNk_DAAAAAAAAAACInxO5rjy5hAKQzSidx2FIQF-FyoexAvamvvJiqNWX-H")
 
@@ -34,8 +36,7 @@ def decrypt(mensagem, senha):
 # ------------ FUNÇÕES DE CRIPTOGRAFIA ------------
 
 def print_ajuda():
-	print(
-"""
+	print("""
 	Exemplos de uso:
 
 	Se um usuário já existente deseja efetuar login
@@ -45,15 +46,17 @@ def print_ajuda():
 	$ python TrabIPPD.py --novousuario <nome do usuario>
 
 	[!] Importante: não é permitido nomes de usuário
-					repetidos.
-""")
+					repetidos.""")
 
 def baixar_lista_usuarios():
-	# baixar a lista de usuarios do dropbox
+	"""Baixa a lista de usuários (usuarios.meta) na a máquina host para extrair
+	a lista de usuários cadastrados"""
 	global dbx
 
 	try:
-		dbx.files_download_to_file("usuarios.meta", "/TrabIPPD/metadados/usuarios.meta")
+		# faz o download do arquivo do dropbox para o arquivo
+		dbx.files_download_to_file("usuarios.meta", 
+			"/TrabIPPD/metadados/usuarios.meta")
 		usuarios = open("usuarios.meta", 'r')
 		usuarios_l = usuarios.read().split(';')
 		usuarios.close()
@@ -63,8 +66,9 @@ def baixar_lista_usuarios():
 		# Lista de usuarios ainda não existe
 		return []
 		
-
 def atualizar_lista_usuarios(lista_usuarios_atualizada):
+	"""Atualiza a lista de usuários no arquivo de metadados no computador 
+	host e sincroniza com o Dropbox"""
 	global dbx
 
 	usuarios = open("usuarios.meta", 'w')
@@ -86,9 +90,8 @@ def atualizar_lista_usuarios(lista_usuarios_atualizada):
 		cprint("[-] Falha na comunicação com o Dropbox.", 'white', 'on_red')
 		return False
 
-
 def baixar_arquivo_senha_usuario(nome):
-	# baixa o arquivo da senha do usuario
+	"""Baixa o arquivo da senha do usuario"""
 	try:
 		fonte = "/TrabIPPD/metadados/{}.key".format(nome)
 		dbx.files_download_to_file("{}.key".format(nome), fonte)
@@ -100,19 +103,24 @@ def baixar_arquivo_senha_usuario(nome):
 def baixar_arquivos(diretorio):
 	global dbx
 	# parte responsável por listar os arquivos e fazer o download para uma pasta local
-	resposta = dbx.files_list_folder("/TrabIPPD/" + diretorio)
+	diretorio_dbx = dbx.files_list_folder("/TrabIPPD/" + diretorio)
 
 	try:
-		for entry in resposta.entries:
-			print("[!] Baixando '" + entry.name + "'...", end='')
-			dbx.files_download_to_file(entry.name, "/TrabIPPD/" + diretorio + "/" + entry.name)
+		for arquivo in diretorio_dbx.entries:
+			print("[!] Baixando '" + arquivo.name + "'...", end='')
+			dbx.files_download_to_file(arquivo.name, 
+				"/TrabIPPD/" + diretorio + "/" + arquivo.name)
 			cprint("OK", "green")
 		return True
-	except:
+	except Exception as e:
 		cprint("[-] Falha no download dos arquivos.", "white", "on_red")
+		print(e)
 		return False
 
 def iniciar_sessao(nome):
+	"""Abre um shell de sessão para o usuário, enquanto esse shell estiver ativo,
+	quando esse shell terminar, os arquivos na máquina do usuário serão sincronizados
+	com a pasta do usuário no dropbox"""
 	if platform.system() == 'Windows':
 		cprint("[*] Abrindo PowerShell para {}".format(nome), 'green')
 		subprocess.run(["powershell"])
@@ -121,32 +129,73 @@ def iniciar_sessao(nome):
 		subprocess.run(["bash"])
 	else:
 		cprint("[-] Plataforma não suportada", "white", "on_red")
-		#exit()
 
-def fazer_backup(diretorio):
-	print("[*] Fazendo files_uploadad dos arquivos...")
-	#após terminar o bash, upa todos os arquivos de volta pro dropbox
-	for entry in os.listdir():
-		try:
-			print("[!] Fazendo o upload de '" + entry + "'...", end='')
-			f = open(entry, 'rb')
-			dbx.files_upload(f.read(), "/TrabIPPD/" + diretorio + "/"+entry, mode=dropbox.files.WriteMode("overwrite"))
-			f.close()
-			os.remove(entry)
-			cprint("OK", "green")
-		except Exception as e:
-			cprint("FALHA", "red")
-			print()
-			cprint("[-] Somente backup de arquivos.", "white", "on_red")
 
-	# feito isso sobe um nível
-	cprint("[+] Saindo do diretório...", "green")
-	os.chdir("..")
+def calcular_deltas(diretorio):
+	global dbx
+	diretorio_dbx = dbx.files_list_folder("/TrabIPPD/" + diretorio).entries
 
-	# e remove o diretório
-	os.rmdir(diretorio)
+	diretorio_dbx = [arq.name for arq in diretorio_dbx]
+	diretorio_host = os.listdir()
+
+	# calcula a diferença entre os itens no dropbox e os arquivos no host
+	delta = list(set(diretorio_dbx) - set(diretorio_host))
+	return delta
+
+def fazer_backup(nome, diretorio):
+	"""Envia todos os arquivos no diretório atual para a pasta do usário"""
+	total = sum(os.path.getsize(f) for f in os.listdir('.') if os.path.isfile(f))
+
+	print("Total de {}".format(size(total)))
+
+	print("[*] Consultando Dropbox sobre o espaço disponível... ", end='')
+	resp = dbx.users_get_space_usage()
+	livre = resp.allocation.get_individual().allocated - resp.used
+	print("Espaço disponível: {}".format(size(livre)))
+
+	if livre > total:
+		# calcula a diferença entre os arquivos no host e os arquivos no dropbox
+		delta = calcular_deltas(diretorio)
+
+		# faz o backup de todos os arquivos que estão no host
+		print("[*] Fazendo upload dos arquivos...")
+		for arquivo in os.listdir():
+			try:
+				print("[!] Fazendo o upload de '" + arquivo + "'...", end='')
+				f = open(arquivo, 'rb')
+				dbx.files_upload(f.read(), "/TrabIPPD/" + diretorio + "/"+arquivo, mode=dropbox.files.WriteMode("overwrite"))
+				f.close()
+				os.remove(arquivo)
+				cprint("OK", "green")
+			except Exception as e:
+				cprint("FALHA", "red")
+				print()
+				cprint("[-] Somente backup de arquivos.", "white", "on_red")
+
+		# remove do dropbox todos os arquivos que não estavam no host
+		print("[!] Resolvendo diferenças...", end='')
+		for arq in delta:
+			dbx.files_delete("/TrabIPPD/" + diretorio + "/" + arq)
+
+		cprint("OK", 'green')
+
+		# sobe para a pasta acima
+		cprint("[+] Saindo do diretório...", "green")
+		os.chdir("..")
+
+		# remove o diretório
+		os.rmdir(diretorio)
+	else:
+		cprint("[-] Não há espaço disponível no Dropbox para fazer o backup de arquivos.", 'white', 'on_red')
+		print("[*] Tente novamente em uma nova sessão.")
+		iniciar_sessao(nome)
+		fazer_backup(nome, diretorio)
+
+
 
 def criar_usuario(nome, usuarios_existentes):
+	"""Cria um novo usuário no sistema dado o nome do novo usuário e a lista de 
+	usuários existentes"""
 	global dbx
 	if nome in usuarios_existentes:
 		cprint("[-] Usuário já existe.", 'white', 'on_red')
@@ -164,13 +213,11 @@ def criar_usuario(nome, usuarios_existentes):
 			dbx.files_upload(senha_crip, 
 				"/TrabIPPD/metadados/{}.key".format(nome), 
 				mode=dropbox.files.WriteMode("overwrite"))
-			print("upload 1")
 
 			# cria um arquivo padrão a todos os usuários
 			dbx.files_upload("Olá".encode(), 
 				"/TrabIPPD/{}/README.md".format(nome), 
 				mode=dropbox.files.WriteMode("overwrite"))
-			print("upload 2")
 
 			os.remove('{}.key'.format(nome))
 			usuarios_existentes.append(nome)
@@ -180,7 +227,9 @@ def criar_usuario(nome, usuarios_existentes):
 			return (False, usuarios_existentes)
 		
 def login(nome, senha):
+	"""Mecanismo que efetua o login no do usuário no sistema"""
 	global dbx
+	# baixa a lista de usuário do dropbox
 	lista_usuarios = baixar_lista_usuarios()
 
 	if nome in lista_usuarios:
@@ -207,6 +256,7 @@ def login(nome, senha):
 		exit()
 
 def sessao():
+	"""Controla a sessão do usuário"""
 	global dbx
 
 	nome = input("[?] Entre com o usuário: ")
@@ -227,7 +277,7 @@ def sessao():
 
 		if baixar_arquivos(diretorio):
 			iniciar_sessao(nome)
-			fazer_backup(diretorio)
+			fazer_backup(nome, diretorio)
 		else:
 			sn = str(input('[!] Deseja tentar novamente? [s, n] '))
 			if sn == 's':
@@ -250,8 +300,11 @@ if __name__ == "__main__":
 			# cria o novo usuário
 			stat, lista_usuarios_atualizada = criar_usuario(nome, lista_usuarios)
 
-			# atualiza a lista de usuários existente
-			atualizar_lista_usuarios(lista_usuarios_atualizada)
+			if stat:
+				# atualiza a lista de usuários existente
+				atualizar_lista_usuarios(lista_usuarios_atualizada)
+			else:
+				exit()
 		else:
 			cprint("[-] Comando não identificado.", "white", "on_red")
 			print_ajuda()
